backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
APScheduler==3.10.4
bcrypt==5.0.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
fastapi==0.124.0
h11==0.16.0
idna==3.11
invoke==2.2.1
markdown-it-py==4.0.0
mdurl==0.1.2
netmiko==4.6.0
ntc_templates==8.1.0
packaging==25.0
paramiko==4.0.0
passlib==1.7.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyNaCl==1.6.1
pynetbox==7.5.0
pyserial==3.5
python-ldap==3.4.5
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
schedule==1.2.2
scp==0.15.0
six==1.17.0
starlette==0.50.0
textfsm==2.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.1
uvicorn==0.38.0
